4.0.0 (#3)
* initial commit for sbs cli tool

* Add files via upload

initialization

* Update requirements.txt

* integrated cli version of sbs v2v

* Update v2v_sbs_converter_cli.sh

* delay time was to long

* Update VR_We_Are_User_Manual.pdf

* Update v2v_sbs_converter_cli.sh

* Update vrweare.py

* limit vres to 2160p

* Enforce sleep / playback fps limit below 50

* Update v2v_sbs_converter_cli.sh

* Update v2v_sbs_converter_cli.sh

* bugfix sageattenstion

* Update forward.sh

* bugfix load failed

* Refactor pipeline to use CloseableQueue

Refactor pipeline to use CloseableQueue for thread.

* Refactor pipeline to use CloseableQueue for thread.

Added error handling and graceful shutdown to worker threads. Enhanced the PipelineContext class with a fatal_error flag.

* Refactor sbsutils.py and add graceful shutdown

* Update v2v_sbs_converter_cli.sh

* Update rating.py

* New Images for sort order selection feature

* file sorting order selection

* scene detection status and request on demand for large videos

* Bugfix seek while play

* bugfix seek crash

* Update rating.py

* snapshot direct without ffmpeg

* File slider preview

* Update rating.py

* stage and workflow changes

* Upgrade to ComfyUI 3.67

* Update VrWeAre-Installer.bat

* bugfix workflow/test

* bugfix

* installation bugfix

* Update VrWeAre-Installer.bat

* Update daemon.bat

* install bugfix

* Update daemon.bat

* installer bugfix

* forward switch

* Forward configuration over app

* deactivate forward by config

* Update forward.sh

* Update VrWeAre-Installer.bat

* preparations 4.0.0

* Update VrWeAre-Installer.bat

* global node name conflict and missing tooltip

* fix and improve workflow

* download URLs fixed.

* Bugfix script and upgrade.

* Add video quality settings to pipeline

add --quality argument for video mode (low, medium, high) to control CRF/CQ values

* Enhance video processing with quality settings

Added video quality parameter to PipelineContext and updated video_worker_thread to handle CRF and CQ settings for video encoding.

* update config validation and preset merging to support video_quality field

* Create install-test.yml

* Update install-test.yml

* Update install-test.yml

* working on runner

* fix runner

* Update install-test.yml

* fix runnter

* Update VrWeAre-Installer.bat

* Update VrWeAre-Installer.bat

* Update VrWeAre-Installer.bat

* runner test activated

* Update VrWeAre-Installer.bat

* fix runner

* Update VrWeAre-Installer.bat

* Update VrWeAre-Installer.bat

* fix runner

* test changes

* Update VrWeAre-Installer.bat

* fix install

* Update VrWeAre-Installer.bat

* i2i sbs cli and v2v quality support

* installer extras

* fix script

* Update install-test.yml

* fix cache and license

* Update install-test.yml

* fix script

* cache test

* add exiftool to runner image.

* Update install-test.yml

* Update install-test.yml

* Update install-test.yml

* Update install-test.yml

* Update install-test.yml

* Update install-test.yml

* WIP

* Update VrWeAre-Installer.bat

* retry

* retry

* add dnd support

* try to print test log

* Update prerequisites.sh

* retry

* change test run

* skip tests on runner

* fix success status

* don't verify

* dnd files dirs and urls

* fixes

* retest cache

* retest

* fix file slider layout

* prepared beta

* Update VrWeAre-Installer.bat

* remove cache

---------

+++ api/prerequisites.sh
@@ -137,6 +137,10 @@ if [ ! -e $CONFIGFILE ] ; then
 	echo "VIDEO_FORMAT=video/h264-mp4">>"$CONFIGFILE"
 	echo "VIDEO_PIXFMT=yuv420p">>"$CONFIGFILE"
 	echo "VIDEO_CRF=17">>"$CONFIGFILE"
+	echo "# --- CLI  (true/false) ---">>"$CONFIGFILE"
+	echo "CLI_ENABLED=true">>"$CONFIGFILE"
+	echo "# --- CLI Video Output quality: low, medium or high ---">>"$CONFIGFILE"
+	echo "VIDEOQUALITYPRESET=high">>"$CONFIGFILE"
 	echo "">>"$CONFIGFILE"
 	
 	echo "# --- scaling config ---">>"$CONFIGFILE"
@@ -422,7 +426,11 @@ cd ../..
 
 # REBUILD FORWARD PIPELINE
 if [ ! -e "$CONFIGPATH"/"autoforward.yaml" ] ; then
-	cp ./custom_nodes/comfyui_stereoscopic/config/default_autoforward.yaml "$CONFIGPATH"/"autoforward.yaml"
+	if [ -e "$CONFIGPATH"/"default_autoforward.yaml" ] ; then
+		cp "$CONFIGPATH"/"default_autoforward.yaml" "$CONFIGPATH"/"autoforward.yaml"	
+	else
+		cp ./custom_nodes/comfyui_stereoscopic/config/default_autoforward.yaml "$CONFIGPATH"/"autoforward.yaml"
+	fi
 fi
 echo -e $"\e[2mRebuild forward rules with \e[36m$CONFIGPATH""/autoforward.yaml\e[0m"
 # Clear forward definitions and rebuild.
@@ -460,8 +468,7 @@ for task in $TASKDIR; do
 	fi
 done
 
-
-# CHECK FOR VERSION UPDATE AND RUN TESTS
+# CHECK FOR VERSION UPDATE FLAGGED BY __init__.py AND RUN TESTS
 if [ -e "custom_nodes/comfyui_stereoscopic/.test/.install" ] ; then
 	./custom_nodes/comfyui_stereoscopic/tests/run_tests.sh || exit 1
 	if [ -e "custom_nodes/comfyui_stereoscopic/.test/.install" ] ; then
